Se eliminan referencias al factor de sesiones consecutivas en varias clases y se actualizan los DTOs para manejar puntuaciones y porcentajes de manera más estructurada. Se implementan cambios en el servicio GeneradorService para optimizar el cálculo de puntuaciones y la gestión de soluciones, mejorando la lógica del generador.

+++ src/main/java/es/iesjandula/reaktor/school_manager_server/dtos/GeneradorInstanciaDto.java
@@ -12,7 +12,7 @@ public class GeneradorInstanciaDto
     private int idGeneradorInstancia ;
 
     /** Puntuación de la instancia del generador */
-    private int puntuacion ;
+    private double puntuacion ;
 
     /** Solución elegida */
     private Boolean solucionElegida ;
